[Prashant] Multithreading is added

+++ EmployeePayrollService/EmployeeModel.cs
@@ -18,7 +18,7 @@ namespace EmployeePayrollService
         public string Address { get; set; }
         public string Department { get; set; }
         public char Gender { get; set; }
-        public decimal BasicPay { get; set; }
+        public double BasicPay { get; set; }
         public double Deductions { get; set; }
         public double TaxablePay { get; set; }
         public double Tax { get; set; }
